Sanitary cleanup: duplicated *dump() + volatile chr + few others
